Fix subcommand help request.
In case of a mapper.Command() call in main, for example `worker queue -h`,
the command is a QueueCmd, but the flag parsing leads to ErrHelp.

So in main, cmdName is queue, but err is ErrHelp. In this case we would get
usage directly from the command, but then exit with status 1 because we would
hit the default case in the switch. This doesn't clean up everything, but moves
in the right direction by exiting cleanly.