Minor help improvements (#110)
* Only show "commands:..." if subcommands exist.

* Fix subcommand help request.

In case of a mapper.Command() call in main, for example `worker queue -h`,
the command is a QueueCmd, but the flag parsing leads to ErrHelp.

So in main, cmdName is queue, but err is ErrHelp. In this case we would get
usage directly from the command, but then exit with status 1 because we would
hit the default case in the switch. This doesn't clean up everything, but moves
in the right direction by exiting cleanly.

* Usage strings to top level descriptions.

* register and run command setup is the same. refactor.

* Remove run command from doc for now. Will remove entirely in another PR.

* Align main help commands.

* Recover from iron_go3 panics to not display stack trace.

* Return err, not nil. Prevents confusion.

* This may make help logic clearer.